Added field price for Carrier

+++ reservations-app/src/main/java/jwachala/project/reservationsapp/BankAccountModel.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class BankAccountModel {
     private String email;
-    private int accountBalance = 0;
+    private double accountBalance = 0;
 
     public BankAccountModel(String email) {
         this.email = email;
